+++ components/Search.js
@@ -2,57 +2,8 @@
 import React, { useState } from 'react';
 import Link from 'next/link';
 import styles from '../styles/Search.module.css';
-
-const MAX_CONTENT_LENGTH = 100;
-
-function searchTrie(trie, searchTerm) {
-  let node = trie;
-  for (const char of searchTerm.split('')) {
-    if (!node[char]) {
-      return [];
-    }
-    node = node[char];
-  }
-
-  return node.words || [];
-}
-
-// TODO: Deduplicate
-const processTitle = title => title.toLowerCase().replaceAll(' ', '-');
-
-function trim(content) {
-  if (content.toString().length <= MAX_CONTENT_LENGTH) {
-    return content;
-  }
-
-  const halfMax = MAX_CONTENT_LENGTH / 2;
-
-  const [start, mid, end] = content;
-
-  // TODO: Better way to do this than using halfMax
-  const trimmedStart =
-    start.length < halfMax
-      ? start
-      : '...' + start.slice(start.length - halfMax);
-
-  const trimmedEnd = end.length < halfMax ? end : end.slice(0, halfMax) + '...';
-
-  return [trimmedStart, mid, trimmedEnd];
-}
-
-function splitOnTerm(content = '', searchTerm = '') {
-  const lineIndex = content.toLowerCase().indexOf(searchTerm.toLowerCase());
-
-  if (lineIndex === -1) {
-    return [content];
-  }
-
-  return trim([
-    content.slice(0, lineIndex),
-    content.slice(lineIndex, lineIndex + searchTerm.length),
-    content.slice(lineIndex + searchTerm.length),
-  ]);
-}
+import { searchTrie, splitOnTerm } from '../utils/search';
+import { processTitle } from '../utils/formatting';
 
 function Search({ searchTerms, trie, episodeMap }) {
   const [searchTerm, setSearchTerm] = useState('');
