+++ styles/Index.module.css
@@ -14,6 +14,6 @@
 
 .episodesHeader {
   display: flex;
-  align-items: end;
+  align-items: flex-end;
   gap: 1.5em;
 }
