+++ utils/formatting.js
@@ -1,0 +1,1 @@
+export const processTitle = title => title.toLowerCase().replaceAll(' ', '-');
